refactor: decouple MeshesInfo from Utils assembly
