+++ unity-renderer/Assets/ABConverter/Utils.cs
@@ -16,6 +16,27 @@ using static DCL.ContentServerUtils;
 
 namespace DCL.ABConverter
 {
+    public static class MeshUtils
+    {
+        public static Bounds BuildMergedBounds(Renderer[] renderers)
+        {
+            Bounds bounds = new Bounds();
+
+            for (int i = 0; i < renderers.Length; i++)
+            {
+                if (renderers[i] == null)
+                    continue;
+
+                if (i == 0)
+                    bounds = renderers[i].GetSafeBounds();
+                else
+                    bounds.Encapsulate(renderers[i].GetSafeBounds());
+            }
+
+            return bounds;
+        }
+    }
+
     public static class PathUtils
     {
         /// <summary>
