+++ unity-renderer/Assets/Scripts/MainScripts/DCL/Helpers/Utils/Utils.cs
@@ -423,24 +423,6 @@ namespace DCL.Helpers
             }
         }
 
-        public static Bounds BuildMergedBounds(Renderer[] renderers)
-        {
-            Bounds bounds = new Bounds();
-
-            for (int i = 0; i < renderers.Length; i++)
-            {
-                if (renderers[i] == null)
-                    continue;
-
-                if (i == 0)
-                    bounds = renderers[i].GetSafeBounds();
-                else
-                    bounds.Encapsulate(renderers[i].GetSafeBounds());
-            }
-
-            return bounds;
-        }
-
         private static int lockedInFrame = -1;
         public static bool LockedThisFrame() => lockedInFrame == Time.frameCount;
